# bos-workspace/backend-nestjs/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/workflows": {
      "post": {
        "operationId": "WorkflowsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkflowDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tạo mới một Quy trình",
        "tags": [
          "Workflow Engine (Cỗ máy Quy trình)"
        ]
      },
      "get": {
        "operationId": "WorkflowsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "example": "id",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy danh sách Quy trình (Đã lọc theo tầm nhìn phân quyền, có phân trang & sắp xếp)",
        "tags": [
          "Workflow Engine (Cỗ máy Quy trình)"
        ]
      }
    },
    "/api/v1/workflows/{id}": {
      "get": {
        "operationId": "WorkflowsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy chi tiết Quy trình kèm các Phiên bản",
        "tags": [
          "Workflow Engine (Cỗ máy Quy trình)"
        ]
      },
      "patch": {
        "operationId": "WorkflowsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkflowDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cập nhật tên/mô tả Quy trình",
        "tags": [
          "Workflow Engine (Cỗ máy Quy trình)"
        ]
      },
      "delete": {
        "operationId": "WorkflowsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Xóa Quy trình",
        "tags": [
          "Workflow Engine (Cỗ máy Quy trình)"
        ]
      }
    },
    "/api/v1/workflows/{id}/versions/{versionId}/clone": {
      "post": {
        "operationId": "WorkflowsController_cloneVersion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Nhân bản một Version thành Version mới",
        "tags": [
          "Workflow Engine (Cỗ máy Quy trình)"
        ]
      }
    },
    "/api/v1/workflows/{id}/versions/{versionId}/status": {
      "patch": {
        "operationId": "WorkflowsController_updateVersionStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cập nhật trạng thái Version",
        "tags": [
          "Workflow Engine (Cỗ máy Quy trình)"
        ]
      }
    },
    "/api/v1/workflows/instances/start": {
      "post": {
        "operationId": "WorkflowsController_startInstance",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInstanceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Khởi chạy luồng quy trình cho 1 bản ghi (Trình ký phiếu)",
        "tags": [
          "Workflow Engine (Cỗ máy Quy trình)"
        ]
      }
    },
    "/api/v1/workflows/instances/{id}/action": {
      "post": {
        "operationId": "WorkflowsController_handleAction",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowActionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Thực hiện Phê duyệt hoặc Từ chối phiếu",
        "tags": [
          "Workflow Engine (Cỗ máy Quy trình)"
        ]
      }
    },
    "/api/v1/workflows/instances/{id}/otp-request": {
      "post": {
        "operationId": "WorkflowsController_requestOtp",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Yêu cầu gửi mã xác thực OTP qua email để ký duyệt",
        "tags": [
          "Workflow Engine (Cỗ máy Quy trình)"
        ]
      }
    },
    "/api/v1/workflows/instances/{id}/logs": {
      "get": {
        "operationId": "WorkflowsController_getInstanceLogs",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy lịch sử Audit Log phê duyệt của lượt chạy",
        "tags": [
          "Workflow Engine (Cỗ máy Quy trình)"
        ]
      }
    },
    "/api/v1/workflows/instances/record/{recordId}/logs": {
      "get": {
        "operationId": "WorkflowsController_getRecordInstanceLogs",
        "parameters": [
          {
            "name": "recordId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy lịch sử Audit Log của lượt chạy mới nhất theo recordId",
        "tags": [
          "Workflow Engine (Cỗ máy Quy trình)"
        ]
      }
    },
    "/api/v1/workflows/instances/record/{recordId}/progress": {
      "get": {
        "operationId": "WorkflowsController_getRecordInstanceProgress",
        "parameters": [
          {
            "name": "recordId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy lộ trình và tiến độ phê duyệt của lượt chạy mới nhất",
        "tags": [
          "Workflow Engine (Cỗ máy Quy trình)"
        ]
      }
    },
...
  },
  "info": {
    "title": "BOS Core Engine API",
    "description": "Tài liệu API cho Nền tảng Quản trị Doanh nghiệp (BOS v2.0)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateWorkflowDto": {
        "type": "object",
        "properties": {
          "entityId": {
            "type": "number",
            "description": "ID của Entity (Biểu mẫu) gắn với quy trình này",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Tên quy trình",
            "example": "Quy trình xét duyệt mua sắm"
          },
          "description": {
            "type": "string",
            "description": "Mô tả chung"
          },
          "visibility": {
            "type": "object",
            "description": "Cấu hình phân quyền tầm nhìn",
            "example": {
              "allowAll": false,
              "allowedDepartments": [
                1
              ]
            }
          }
        },
        "required": [
          "entityId",
          "name"
        ]
      },
      "UpdateWorkflowDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Tên quy trình",
            "example": "Quy trình xét duyệt mua sắm"
          },
          "description": {
            "type": "string",
            "description": "Mô tả chung"
          },
          "visibility": {
            "type": "object",
            "description": "Cấu hình phân quyền tầm nhìn",
            "example": {
              "allowAll": false,
              "allowedDepartments": [
                1
              ]
            }
          }
        }
      },
      "CreateInstanceDto": {
        "type": "object",
        "properties": {
          "recordId": {
            "type": "number",
            "description": "ID của Bản ghi dữ liệu (Form đã submit)",
            "example": 1
          },
          "versionId": {
            "type": "number",
            "description": "ID của Workflow Version muốn áp dụng",
            "example": 2
          },
          "nextAssigneeId": {
            "type": "number",
            "description": "ID của người duyệt tiếp theo được chỉ định động",
            "example": 10
          }
        },
        "required": [
          "recordId",
          "versionId"
        ]
      },
      "WorkflowActionDto": {
        "type": "object",
        "properties": {
          "transitionId": {
            "type": "number",
            "description": "ID của đường nối (Nút bấm) mà người dùng chọn kích hoạt",
            "example": 1
          },
          "comment": {
            "type": "string",
            "description": "Ý kiến phê duyệt",
            "example": "Đồng ý chuyển tiếp hồ sơ."
          },
          "signatureData": {
            "type": "string",
            "description": "Mã chữ ký điện tử hoặc chuỗi nét vẽ Base64 (Nếu nút yêu cầu ký duyệt)",
            "example": "SIGNED_BY_ADMIN_MD5_789abc"
          },
          "otpCode": {
            "type": "string",
            "description": "Mã xác thực OTP (Nếu nút yêu cầu ký duyệt)",
            "example": "123456"
          },
          "stampData": {
            "type": "string",
            "description": "Dữ liệu ảnh con dấu Base64 đi kèm (Nếu có)",
            "example": "data:image/png;base64,..."
          },
          "signatureLayout": {
            "type": "string",
            "description": "Bố cục hiển thị thông tin chữ ký: dọc hoặc ngang",
            "enum": [
              "vertical",
              "horizontal"
            ],
            "example": "vertical"
          },
          "showSignerName": {
            "type": "boolean",
            "description": "Hiển thị họ tên người ký",
            "example": true
          },
          "showSignerRole": {
            "type": "boolean",
            "description": "Hiển thị vai trò/chức danh người ký",
            "example": true
          },
          "showSignerDept": {
            "type": "boolean",
            "description": "Hiển thị phòng ban người ký",
            "example": true
          },
          "showSigningTime": {
            "type": "boolean",
            "description": "Hiển thị thời gian thực hiện ký",
            "example": true
          },
          "nextAssigneeId": {
            "type": "number",
            "description": "ID của người duyệt tiếp theo được chọn động",
            "example": 101
          }
        },
        "required": [
          "transitionId"
        ]
      },
...
    }
  }
}